backend/pyproject.toml
[project]
name = "pikoshi"
version = "0.0.0"
description = "Pikoshi: The Photo Sharing Application With Privacy In Mind By Default 🐱"
authors = [
  { name = "tomit4", email = "mosssap@gmail.com" },
  { name = "paulcap510", email = "paul.capob@gmail.com" },
]
dependencies = [
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.32",
    "alembic>=1.13.2",
    "httpx>=0.27.0",
    "pydantic[email]>=2.8.2",
    "resend>=2.4.0",
    "redis>=5.0.8",
    "pyjwt>=2.9.0",
    "boto3>=1.35.10",
    "python-multipart>=0.0.9",
    "pillow>=10.4.0",
    "asyncpg>=0.29.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "black>=24.8.0",
    "isort>=5.13.2",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pikoshi"]

[tool.rye.scripts]
dev = "uvicorn pikoshi.main:app --host :: --port 8000 --reload"
upgrade = "alembic --config ./src/pikoshi/alembic.ini upgrade"
downgrade = "alembic --config ./src/pikoshi/alembic.ini downgrade"
generate = "alembic --config ./src/pikoshi/alembic.ini revision --autogenerate -m"

[project.scripts]
start = "pikoshi.main:main"
